move walker/filter.hh -> filters/...
