+++ dune/xt/grid/filters.hh
@@ -1,0 +1,17 @@
+// This file is part of the dune-xt-grid project:
+//   https://github.com/dune-community/dune-xt-grid
+// Copyright 2009-2018 dune-xt-grid developers and contributors. All rights reserved.
+// License: Dual licensed as BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
+//      or  GPL-2.0+ (http://opensource.org/licenses/gpl-license)
+//          with "runtime exception" (http://www.dune-project.org/license.html)
+// Authors:
+//   Felix Schindler (2018)
+
+#ifndef DUNE_XT_GRID_FILTERS_HH
+#define DUNE_XT_GRID_FILTERS_HH
+
+#include "filters/base.hh"
+#include "filters/element.hh"
+#include "filters/intersection.hh"
+
+#endif // DUNE_XT_GRID_FILTERS_HH
